Move move Route facade methods to Routes facade
