+++ packages/framework/src/Facades/Route.php
@@ -4,11 +4,6 @@ declare(strict_types=1);
 
 namespace Hyde\Facades;
 
-use Hyde\Hyde;
-use Hyde\Foundation\Facades\Routes;
-use Hyde\Foundation\Kernel\RouteCollection;
-use Hyde\Support\Models\RouteKey;
-
 /**
  * Provides an easy way to access the Hyde pseudo-router.
  *
@@ -16,31 +11,5 @@ use Hyde\Support\Models\RouteKey;
  */
 class Route
 {
-    public static function exists(string $routeKey): bool
-    {
-        return Routes::has(RouteKey::normalize($routeKey));
-    }
-
-    public static function get(string $routeKey): ?\Hyde\Support\Models\Route
-    {
-        return Routes::get(RouteKey::normalize($routeKey));
-    }
-
-    /** @throws \Hyde\Framework\Exceptions\RouteNotFoundException */
-    public static function getOrFail(string $routeKey): \Hyde\Support\Models\Route
-    {
-        return Routes::getRoute(RouteKey::normalize($routeKey));
-    }
-
-    /** @return \Hyde\Foundation\Kernel\RouteCollection<\Hyde\Support\Models\Route> */
-    public static function all(): RouteCollection
-    {
-        return Routes::getRoutes();
-    }
-
-    /** Get the current route for the page being rendered. */
-    public static function current(): ?\Hyde\Support\Models\Route
-    {
-        return Hyde::currentRoute();
-    }
+    //
 }
